pyproject.toml
[tool.poetry]
name = "e-cup-ozon"
version = "0.1.0"
description = "Competition experimentation framework (data splits, training, evaluation, artifacts)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
hydra-core = "1.3.2"
omegaconf = "2.3.0"
polars = ">=0.20.0"
pandas = ">=2.2.0"
numpy = ">=1.26.0"
scikit-learn = ">=1.4.0"
lightgbm = ">=4.3.0"
mlflow = ">=2.14.0"
dvc = ">=3.51.0"
pyarrow = ">=14.0.0"
pydantic = ">=2.7.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.0"

[tool.poetry.group.recbole.dependencies]
recbole = ">=1.2.0"
torch = ">=2.2.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
